+++ freppledb/output/commands.py
@@ -362,7 +362,7 @@ class ExportOperationPlans(PlanTask):
 
     @classmethod
     def getWeight(cls, **kwargs):
-        if "supply" in os.environ:
+        if "supply" in os.environ or kwargs.get("exportstatic", False):
             return 1
         else:
             return -1